+++ packages/sitebuilder/src/setup/SetupCard.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { __ } from '@wordpress/i18n';
 import { Chip, SetupCard as WmeSetupCard, SetupCardHeader, SetupCardContent } from '@moderntribe/wme-ui';
 import { WatchLater, CheckCircle } from '@mui/icons-material';