+++ packages/sitebuilder/src/wizards/WizardContent.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Box } from '@mui/material';
 import type { BoxProps } from '@mui/material/Box';
 import { styled } from '@mui/material/styles';