+++ packages/sitebuilder/src/wizards/WizardHeader.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { WizardHeader as WmeWizardHeader, Logo, ExitButton } from '@moderntribe/wme-ui';
 import { useLocation } from 'react-router-dom';
 import { __ } from '@wordpress/i18n';