+++ packages/sitebuilder/src/wizards/first-time-configuration/screens/Complete.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { WizardSectionTitle, Button } from '@moderntribe/wme-ui';
 import { Box, Typography } from '@mui/material';
 import { useFirstTimeConfiguration } from '@sb/hooks';