+++ packages/sitebuilder/src/wizards/go-live/data/constants.ts
@@ -4,6 +4,7 @@ export const GoLiveStringData = {
 	goLiveProviderText: {
 		getDomain: __('Get My Domain', 'moderntribe-sitebuilder'),
 		haveDomain: __('I\'ve Got My Domain', 'moderntribe-sitebuilder'),
+    checkout: __('Checkout', 'moderntribe-sitebuilder'),
 		continueStr: __('Continue', 'moderntribe-sitebuilder'),
 		errorMessage: __('There was an error while updating your site\'s domain.', 'moderntribe-sitebuilder'),
 		errorMessageVerification: __('There was an error while verifying your site\'s domain.', 'moderntribe-sitebuilder'),
@@ -13,21 +14,23 @@ export const GoLiveStringData = {
 	},
 	start: {
 		screenTitle1: __('Set your site domain', 'moderntribe-sitebuilder'),
-		screenTitle2: __('Own your perfect domain with Nexcess.', 'moderntribe-sitebuilder'),
+		screenTitle2: __('Find the perfect domain.', 'moderntribe-sitebuilder'),
 		screenDescription1: __('Let\'s get your site set up with the perfect domain. If you don\'t already have one, don\'t worry - we can help you out.', 'moderntribe-sitebuilder'),
-		screenDescription2: __('Nexcess helps you find, create and own your perfect domain. It\'s so easy, even we could do it. Once you\'ve got your domain, come back here and select "I\'ve Got My Domain".', 'moderntribe-sitebuilder'),
+		screenDescription2: __('Enter the domain name you want your store to have, and we\'ll see if it\'s available. If it isn\'t, we\'ll make suggestions about ones that are.', 'moderntribe-sitebuilder'),
 		actionLabel: __('Have you purchased your custom domain?', 'moderntribe-sitebuilder'),
 		actionTitle1: __('Yes! I have my own custom domain.', 'moderntribe-sitebuilder'),
 		actionTitle2: __('No, I need one!', 'moderntribe-sitebuilder'),
-		actionContent1: __('If you have your own domain already then you\'re ready to connect and go live.', 'moderntribe-sitebuilder'),
-		actionContent2: __('Nexcess will help you find and own your perfect domain name.', 'moderntribe-sitebuilder')
+		actionContent1: __('If you have your own domain already then you\'re ready to connect and go live.', 'moderntribe-sitebuilder),
+		actionContent2: __('Nexcess will help you find and own your perfect domain name.', 'moderntribe-sitebuilder'),
+		defaultError: __('Something went wrong, please try again', 'moderntribe-sitebuilder')
 	},
 	verifyDomain: {
 		screenTitle: __('Verify your domain', 'moderntribe-sitebuilder'),
 		screenDescription: __('We\'ll verify your domain and get it ready to use. This can take a while, but don\'t worry - we\'ll walk you through it.', 'moderntribe-sitebuilder'),
+    screenNotice: __('If you\'re using a subdomain (ex: store.example.com), please stop here and reach out to support at', 'moderntribe-sitebuilder'),
 		goLiveLabelText: __('Enter the domain you want to connect', 'moderntribe-sitebuilder'),
 		goLivePlaceholderText: __('yourdomain.com', 'moderntribe-sitebuilder'),
-		errorDomainFormat: __('The domain entered does not appear to be a valid format.', 'moderntribe-sitebuilder'),
+		errorDomainFormat: __('The domain entered does not appear to be a valid format.', 'moderntribe-sitebuilder')
 	},
 	updateSiteUrl: {
 		screenTitle: __('You\'re ready to go live with', 'moderntribe-sitebuilder'),
@@ -36,6 +39,14 @@ export const GoLiveStringData = {
 		descriptionProcessing: __('This can take a few minutes. You can log out now, or we can will just log out upon completion.', 'moderntribe-sitebuilder'),
 		loginUrlLabelText: __('Copy Your New Login URL', 'moderntribe-sitebuilder'),
 		loginUrlHelperText: __('Copy and save this URL. Once your domain is connected you\'ll need it to login.', 'moderntribe-sitebuilder')
+	},
+	connectWithNexcess: {
+		screenTitle: __('Good choice! We’ll need to connect to your Nexcess account.', 'nexcess-mapps'),
+		screenDescription: __('We\'ll make sure your chosen domain is managed by your Nexcess account. Once connected, we\'ll navigate you back here to complete the process.', 'nexcess-mapps'),
+	},
+	claimYourDomain: {
+		screenTitle: __('is all yours!', 'nexcess-mapps'),
+		screenDescription: __('While a new domain is often ready for use within an hour, it can take up to 8 hours to completely process.', 'nexcess-mapps'),
 	},
 	errorDomainGeneral: {
 		accountContent: __('It looks like we\'re having trouble connecting your domain.', 'moderntribe-sitebuilder'),
@@ -74,5 +85,10 @@ export const GoLiveStringData = {
 		message2: __('If anything does go wrong, please reach out to support and we\'ll get you up and running!', 'moderntribe-sitebuilder'),
 		continueButton: __('Continue without verification', 'moderntribe-sitebuilder'),
 		nevermind: __('Do not skip verification', 'moderntribe-sitebuilder')
+	},
+	domainItems: {
+		taken: __('Taken', 'nexcess-mapps'),
+		selected: __('Selected', 'nexcess-mapps'),
+		available: __('Available', 'nexcess-mapps')
 	}
 };