+++ packages/storebuilder/src/wizards/WizardHeader.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { WizardHeader as WmeWizardHeader, Logo, ExitButton } from '@moderntribe/wme-ui';
 import { __ } from '@wordpress/i18n';
 import { useWizard } from '@store/hooks';