+++ packages/storebuilder/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @moderntribe/storebuilder
 
+## 1.5.0
+
+### Minor Changes
+
+- 538bda1: Developer experience updates
+
+### Patch Changes
+
+- Updated dependencies [538bda1]
+- Updated dependencies [5c6be2a]
+  - @moderntribe/wme-ui@2.3.0
+  - @moderntribe/wme-utils@0.2.0
+
 ## 1.4.2
 
 ### Patch Changes